+++ src/handlers/sourcesHandlers.ts
@@ -1,5 +1,6 @@
 import { HTTP_STATUS } from '../constants';
 import {
+	createPermissionGrant,
 	getDatabase,
 	createSource,
 	getLimitValues,
@@ -11,6 +12,7 @@ import {
 	removeSource,
 } from '../database';
 import {
+	getSelfManageGrantPartial,
 	isAuthContext,
 	isId,
 	isWritableSource,
@@ -31,16 +33,17 @@ const postSource = async (req: Request, res: Response): Promise<void> => {
 		throw new FailedMiddlewareError('Unexpected lack of auth context.');
 	}
 	const db = getDatabase();
-	if (!isWritableSource(req.body)) {
+	const body = req.body as unknown;
+	if (!isWritableSource(body)) {
 		throw new InputValidationError(
 			'Invalid request body.',
 			isWritableSource.errors ?? [],
 		);
 	}
 	if (
-		'funderShortCode' in req.body &&
+		'funderShortCode' in body &&
 		!(await hasFunderPermission(db, req, {
-			funderShortCode: req.body.funderShortCode,
+			funderShortCode: body.funderShortCode,
 			permission: PermissionGrantVerb.CREATE,
 			scope: PermissionGrantEntityType.SOURCE,
 		}))
@@ -50,9 +53,9 @@ const postSource = async (req: Request, res: Response): Promise<void> => {
 		);
 	}
 	if (
-		'dataProviderShortCode' in req.body &&
+		'dataProviderShortCode' in body &&
 		!(await hasDataProviderPermission(db, req, {
-			dataProviderShortCode: req.body.dataProviderShortCode,
+			dataProviderShortCode: body.dataProviderShortCode,
 			permission: PermissionGrantVerb.CREATE,
 			scope: PermissionGrantEntityType.SOURCE,
 		}))
@@ -62,9 +65,9 @@ const postSource = async (req: Request, res: Response): Promise<void> => {
 		);
 	}
 	if (
-		'changemakerId' in req.body &&
+		'changemakerId' in body &&
 		!(await hasChangemakerPermission(db, req, {
-			changemakerId: req.body.changemakerId,
+			changemakerId: body.changemakerId,
 			permission: PermissionGrantVerb.CREATE,
 			scope: PermissionGrantEntityType.SOURCE,
 		}))
@@ -77,11 +80,19 @@ const postSource = async (req: Request, res: Response): Promise<void> => {
 	// Normally we try to avoid passing the body directly vs extracting the values and passing them.
 	// Because because writableSource is a union type it is hard to extract the values directly without
 	// losing type context that the union provided.
-	const source = await createSource(db, req, req.body);
+	const committedSource = await db.transaction(async (txDb) => {
+		const source = await createSource(txDb, req, body);
+		await createPermissionGrant(txDb, req, {
+			...getSelfManageGrantPartial(req),
+			contextEntityType: PermissionGrantEntityType.SOURCE,
+			sourceId: source.id,
+		});
+		return source;
+	});
 	res
 		.status(HTTP_STATUS.SUCCESSFUL.CREATED)
 		.contentType('application/json')
-		.send(source);
+		.send(committedSource);
 };
 
 const getSources = async (req: Request, res: Response): Promise<void> => {